requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
rich>=13.7.0
flask>=3.0.0
gunicorn>=22.0.0
